+++ src/util/__tests__/createServerRootMixin.test.js
@@ -1,6 +1,8 @@
 import Vue from 'vue';
 import { mount } from '@vue/test-utils';
 import _renderToString from 'vue-server-renderer/basic';
+import Router from 'vue-router';
+import Vuex from 'vuex';
 import { createServerRootMixin } from '../createServerRootMixin';
 import InstantSearchSsr from '../../components/InstantSearchSsr';
 import Configure from '../../components/Configure';
@@ -213,6 +215,310 @@ Array [
 ]
 `);
     });
+
+    it('forwards router', async () => {
+      const searchClient = createFakeClient();
+
+      const router = new Router({});
+
+      // there are two renders of App, each with an assertion
+      expect.assertions(2);
+
+      const App = Vue.component('App', {
+        mixins: [
+          forceIsServerMixin,
+          createServerRootMixin({
+            searchClient,
+            indexName: 'hello',
+          }),
+        ],
+        data() {
+          expect(this.$router).toBe(router);
+          return {};
+        },
+        render(h) {
+          return h(InstantSearchSsr, {}, [
+            h(Configure, {
+              attrs: {
+                hitsPerPage: 100,
+              },
+            }),
+            h(SearchBox),
+          ]);
+        },
+        serverPrefetch() {
+          return this.instantsearch.findResultsState(this);
+        },
+      });
+
+      Vue.use(Router);
+
+      const wrapper = new Vue({
+        mixins: [forceIsServerMixin],
+        router,
+        render(h) {
+          return h(App);
+        },
+      });
+
+      await renderToString(wrapper);
+    });
+
+    it('forwards vuex', async () => {
+      const searchClient = createFakeClient();
+
+      Vue.use(Vuex);
+
+      const store = new Vuex.Store();
+
+      // there are two renders of App, each with an assertion
+      expect.assertions(2);
+
+      const App = Vue.component('App', {
+        mixins: [
+          forceIsServerMixin,
+          createServerRootMixin({
+            searchClient,
+            indexName: 'hello',
+          }),
+        ],
+        data() {
+          expect(this.$store).toBe(store);
+          return {};
+        },
+        render(h) {
+          return h(InstantSearchSsr, {}, [
+            h(Configure, {
+              attrs: {
+                hitsPerPage: 100,
+              },
+            }),
+            h(SearchBox),
+          ]);
+        },
+        serverPrefetch() {
+          return this.instantsearch.findResultsState(this);
+        },
+      });
+
+      const wrapper = new Vue({
+        mixins: [forceIsServerMixin],
+        store,
+        render(h) {
+          return h(App);
+        },
+      });
+
+      await renderToString(wrapper);
+    });
+
+    it('forwards props', async () => {
+      const searchClient = createFakeClient();
+
+      // there are two renders of App, each with an assertion
+      expect.assertions(2);
+
+      const someProp = { data: Math.random() };
+
+      const App = Vue.component('App', {
+        mixins: [
+          forceIsServerMixin,
+          createServerRootMixin({
+            searchClient,
+            indexName: 'hello',
+          }),
+        ],
+        props: {
+          someProp: {
+            required: true,
+            type: Object,
+            validator(value) {
+              expect(value).toBe(someProp);
+              return value === someProp;
+            },
+          },
+        },
+        render(h) {
+          return h(InstantSearchSsr, {}, [
+            h(Configure, {
+              attrs: {
+                hitsPerPage: 100,
+              },
+            }),
+            h(SearchBox),
+          ]);
+        },
+        serverPrefetch() {
+          return this.instantsearch.findResultsState(this);
+        },
+      });
+
+      const wrapper = new Vue({
+        mixins: [forceIsServerMixin],
+        render(h) {
+          return h(App, { props: { someProp } });
+        },
+      });
+
+      await renderToString(wrapper);
+    });
+
+    it('forwards slots', async done => {
+      const searchClient = createFakeClient();
+
+      expect.assertions(2);
+
+      const App = Vue.component('App', {
+        mixins: [
+          forceIsServerMixin,
+          createServerRootMixin({
+            searchClient,
+            indexName: 'hello',
+          }),
+        ],
+        render(h) {
+          return h(InstantSearchSsr, {}, this.$slots.default);
+        },
+        serverPrefetch() {
+          return (
+            this.instantsearch
+              .findResultsState(this)
+              .then(res => {
+                expect(
+                  this.instantsearch.mainIndex.getWidgets().map(w => w.$$type)
+                ).toEqual(['ais.configure']);
+
+                expect(res.hello._state.hitsPerPage).toBe(100);
+              })
+              // jest throws an error we need to catch, since stuck in the flow
+              .catch(e => {
+                done.fail(e);
+              })
+          );
+        },
+      });
+
+      const wrapper = new Vue({
+        mixins: [forceIsServerMixin],
+        render(h) {
+          return h(App, [
+            h('template', { slot: 'default' }, [
+              h(Configure, {
+                attrs: {
+                  hitsPerPage: 100,
+                },
+              }),
+            ]),
+          ]);
+        },
+      });
+
+      await renderToString(wrapper);
+      done();
+    });
+
+    // TODO: forwarding of scoped slots doesn't yet work.
+    it.skip('forwards scoped slots', async done => {
+      const searchClient = createFakeClient();
+
+      expect.assertions(2);
+
+      const App = Vue.component('App', {
+        mixins: [
+          forceIsServerMixin,
+          createServerRootMixin({
+            searchClient,
+            indexName: 'hello',
+          }),
+        ],
+        render(h) {
+          return h(InstantSearchSsr, {}, [
+            this.$scopedSlots.default({ test: true }),
+          ]);
+        },
+        serverPrefetch() {
+          return (
+            this.instantsearch
+              .findResultsState(this)
+              .then(res => {
+                expect(
+                  this.instantsearch.mainIndex.getWidgets().map(w => w.$$type)
+                ).toEqual(['ais.configure']);
+
+                expect(res.hello._state.hitsPerPage).toBe(100);
+              })
+              // jest throws an error we need to catch, since stuck in the flow
+              .catch(e => {
+                done.fail(e);
+              })
+          );
+        },
+      });
+
+      const wrapper = new Vue({
+        mixins: [forceIsServerMixin],
+        render(h) {
+          return h(App, {
+            scopedSlots: {
+              default({ test }) {
+                if (test) {
+                  return h(Configure, {
+                    attrs: {
+                      hitsPerPage: 100,
+                    },
+                  });
+                }
+                return null;
+              },
+            },
+          });
+        },
+      });
+
+      await renderToString(wrapper);
+      done();
+    });
+
+    it('forwards root', async () => {
+      const searchClient = createFakeClient();
+
+      // there are two renders of App, each with an assertion
+      expect.assertions(2);
+
+      const App = Vue.component('App', {
+        mixins: [
+          forceIsServerMixin,
+          createServerRootMixin({
+            searchClient,
+            indexName: 'hello',
+          }),
+        ],
+        render(h) {
+          expect(this.$root).toBe(wrapper);
+
+          return h(InstantSearchSsr, {}, [
+            h(Configure, {
+              attrs: {
+                hitsPerPage: 100,
+              },
+            }),
+            h(SearchBox),
+          ]);
+        },
+        serverPrefetch() {
+          return this.instantsearch.findResultsState(this);
+        },
+      });
+
+      const wrapper = new Vue({
+        mixins: [forceIsServerMixin],
+        render(h) {
+          return h(App);
+        },
+      });
+
+      await renderToString(wrapper);
+    });
   });
 
   describe('hydrate', () => {
@@ -412,35 +718,35 @@ Array [
 
       expect(renderArgs).toMatchInlineSnapshot(
         {
-          helper: expect.any(Object),
-          results: expect.any(Object),
+          helper: expect.anything(),
+          results: expect.anything(),
           scopedResults: expect.arrayContaining([
             expect.objectContaining({
-              helper: expect.any(Object),
+              helper: expect.anything(),
               indexId: expect.any(String),
-              results: expect.any(Object),
+              results: expect.anything(),
             }),
           ]),
-          state: expect.any(Object),
-          instantSearchInstance: expect.any(Object),
+          state: expect.anything(),
+          instantSearchInstance: expect.anything(),
         },
         `
 Object {
   "createURL": [Function],
-  "helper": Any<Object>,
-  "instantSearchInstance": Any<Object>,
-  "results": Any<Object>,
+  "helper": Anything,
+  "instantSearchInstance": Anything,
+  "results": Anything,
   "scopedResults": ArrayContaining [
     ObjectContaining {
-      "helper": Any<Object>,
+      "helper": Anything,
       "indexId": Any<String>,
-      "results": Any<Object>,
+      "results": Anything,
     },
   ],
   "searchMetadata": Object {
     "isSearchStalled": false,
   },
-  "state": Any<Object>,
+  "state": Anything,
   "templatesConfig": Object {},
 }
 `